+++ src/lib.rs
@@ -2,4 +2,5 @@
 
 #![forbid(unsafe_code)]
 
+pub mod decode;
 pub mod frame;